requirements.txt
fastapi==0.110.0
fastapi-cache2==0.2.1
fastapi-health==0.4.0
uvicorn==0.28.0
requests==2.31.0
pandas==2.2.1
sqlalchemy==2.0.28
clickhouse-sqlalchemy==0.3.1
psycopg2==2.9.9
asgi_correlation_id==4.3.1
python-dotenv==1.0.1